Refactoring sourcecode split tool actions in separate files xmltools.cpp is now the actual plugin code
