+++ Debug.h
@@ -1,0 +1,6 @@
+#include "StdAfx.h"
+
+void createDebugDlg();
+void showDebugDlg();
+void dbg(CStringW line);
+void dbgln(CStringW line);
