+++ Notepad_plus_msgs.h
@@ -32,7 +32,7 @@
 #include <windows.h>
 #include <tchar.h>
 
-enum LangType {
+enum class LangType {
   L_TEXT, L_PHP, L_C, L_CPP, L_CS, L_OBJC, L_JAVA, L_RC, \
   L_HTML, L_XML, L_MAKEFILE, L_PASCAL, L_BATCH, L_INI, L_ASCII, L_USER, \
   L_ASP, L_SQL, L_VB, L_JS, L_CSS, L_PERL, L_PYTHON, L_LUA, \
